Rename stub.py to common.py
