+++ salt/utils/scm/common.py
